Fix migrations
And change percentage to f64 float

+++ server/graphql/demographic/src/mutations/update_indicator.rs
@@ -51,7 +51,7 @@ pub struct UpdateDemographicIndicatorInput {
     pub name: Option<String>,
     pub base_year: Option<i32>,
     pub base_population: Option<i32>,
-    pub population_percentage: Option<f32>,
+    pub population_percentage: Option<f64>,
     pub year_1_projection: Option<i32>,
     pub year_2_projection: Option<i32>,
     pub year_3_projection: Option<i32>,
